chore: refactor constants and app js
