+++ package.json
@@ -14,6 +14,7 @@
     "glsl": "^0.0.1",
     "glsl-noise": "0.0.0",
     "glslify": "^7.1.1",
+    "leva": "^0.9.35",
     "nanoid": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
