+++ src/App.js
@@ -1,66 +1,22 @@
-import { Physics } from "@react-three/cannon";
-import { Sky } from "@react-three/drei";
-import { Canvas } from "@react-three/fiber";
-import { FPV } from "./components/FPV";
-import { TextureSelector } from "./components/TextureSelector";
-import { Menu } from "./components/Menu";
-import { Help } from "./components/Help";
-import { Scene } from "./components/Scene";
-
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import { useStore } from "./hooks/useStore";
-import settings from "./devOnline";
-import { OrbitControls } from "@react-three/drei";
-import { MakeOnlineConnection } from "./components/multiplayercomps/MakeOnlineConnection";
-import { LoadingWorldPage } from "./components/LoadingWorldPage";
-import LowerControlStrip from "./hooks/LowerControlStrip";
-import TitleConfig from "./components/TitleConfig";
+import settings from "./constants";
+import { MakeOnlineConnection } from "./components/multiplayerComponents/MakeOnlineConnection";
+import TitleScreen from "./components/UIComponents/TitleScreen";
+import CoreGame from "./components/CoreGame";
 
 function App() {
-  const movewithJOY_BOOL = true
   const [establishedConn] = useStore((state) => [state.establishedConn]); //begining of online play
-  const activeTextureREF = useRef("dirt");
 
   const [playerConfigReady,setPCR] = useState(false)
-
-  const chunksmadecounter = useRef({ loaddone: false, track: {count:0,max:settings.worldSettings.worldSize**2} });
-  const loadingscreenhtml = useRef();
-  const [initStatus, setInitStatus] = useState({
-    buildWorkers: 0,
-    initWorkers: 0,
-    initWorld: 0,
-  });
-
-  let moveBools = useRef({
-    moveBackward: false,
-    moveForward: false,
-    moveLeft: false,
-    moveRight: false,
-    jump: false,
-    moveQuick: false,
-    moveQuickTC: 0, //tap count
-    moveQuickTT: 0, //tap time
-    jumpTT: 0,
-    jumpTC: 0,
-    camUp: false,
-    camDown: false,
-    camLeft: false,
-    camRight: false,
-    camCenter: false,
-    camCenterTT: 0,
-    camCenterTC: 0,
-  });
-
-  function updateInitStatus(obj) {
-    setInitStatus({ ...initStatus.current, ...obj });
-  }
 
   function gettingWorldLoadScreen() {
     //this is meant to be a place holder for a potential loading screen as we generate enough of the world before the player.
     //also a waiting screen for when some one is waiting to connect to online server
+    //@TODO: MakeOnlineConnection should be a hook
     return (
       <>
-        <div>gettingWorld</div>
+        <div>Getting World</div>
         <MakeOnlineConnection />
       </>
     );
@@ -72,54 +28,7 @@ function App() {
     setPCR(true)
   }
 
-  function goToGame() {
-    if(playerConfigReady){
-      return (
-        <>
-          <LoadingWorldPage buildWorkers={initStatus.buildWorkers} chunksmadecounter={chunksmadecounter} myRef={loadingscreenhtml} />
-          <Canvas>
-            {settings.hideSky ? <></> : <Sky name={"skyMesh"} sunPosition={[100, 100, 20]} />}
-            <ambientLight intensity={0.5} />
-
-  
-            <Physics>
-              <Scene
-                activeTextureREF={activeTextureREF}
-                updateInitStatus={updateInitStatus}
-                initStatus={initStatus}
-                chunksmadecounter={chunksmadecounter}
-                moveBools={moveBools}
-                movewithJOY_BOOL={movewithJOY_BOOL}
-              />
-            </Physics>
-  
-            {settings.useOrbitals ? <OrbitControls /> : <></>}
-            <axesHelper name={"axesHelper"} scale={10} />
-          </Canvas>
-          {settings.ignoreCameraFollowPlayer ? <></> : <div className="cursor centered absolute">+</div>}
-          {settings.hideUIContent ? (
-            <></>
-          ) : (
-            <>
-              <Menu />
-              <Help />
-            </>
-          )}
-  
-          {settings.hideTextSelect ? <></> : <TextureSelector activeTextureREF={activeTextureREF} />}
-          {settings.movewithJOY_BOOL?<LowerControlStrip moveBools={moveBools} />:<></>}
-        </>
-      );
-    }else{
-      return(
-        <>
-        <TitleConfig playerGivenGameSettings={playerGivenGameSettings}/>
-        </>
-      )
-    }
-  }
-
-  return establishedConn ? goToGame() : gettingWorldLoadScreen();
+  return establishedConn ? playerConfigReady ? <CoreGame /> : <TitleScreen playerGivenGameSettings={playerGivenGameSettings}/> : gettingWorldLoadScreen();
 }
 
 export default App;
