+++ src/components/playerComponents/BasicPlayer.js
